Core service moved from services folder to modules folder
