+++ src/src/userauth/userauth.module.ts
@@ -8,7 +8,7 @@ import { KeycloakModule } from 'src/services/keycloak/keycloak.module';
 import { UserModule } from 'src/user/user.module';
 import { AuthService } from 'src/modules/auth/auth.service';
 import { Method } from '../common/method/method';
-import { AcknowledgementModule } from '../services/acknowledgement/acknowledgement.module';
+import { AcknowledgementModule } from '../modules/acknowledgement/acknowledgement.module';
 
 @Module({
 	imports: [
